Configure tmux for multi-client session sharing
I create the main tmux session, and immediately detach it. I then create
sessions and tmux windows as needed and connect them all to the master
session. Because the master session is already in the background, I can
rebind `C-b d' to kill the current session, handily avoiding the many
useless detached tmux sessions that this master session business usually
brings with it.